security updated

+++ Coders-Platform-Backend/src/main/java/com/javafest/DiffDeptStormers/service/MongoRepoService.java
@@ -38,31 +38,30 @@ public class MongoRepoService {
                 .append("repoName", repository.getRepoName())
                 .append("repoDescription", repository.getRepoDescription())
                 .append("repoTopicTags", repository.getRepoTopicTags())
-                .append("isPublic", repository.isPublic());
+                .append("isPublic", repository.isPublic())
+                .append("email", repository.getEmail());
     }
 
     private Document convertFileToDocument(File file) {
         return new Document()
                 .append("repoId", file.getRepoId())
                 .append("fileName", file.getFileName())
-                .append("fileContent", file.getFileContent());
+                .append("fileContent", file.getFileContent())
+                .append("email", file.getEmail());
     }
 
-    public Repository saveRepository(Repository repository) {
+    public Repository saveRepository(Repository repository, String email) {
+    	repository.setEmail(email);
         Document repoDoc = convertRepositoryToDocument(repository);
         MongoCollection<Document> repoCollection = getRepoCollection();
         repoCollection.insertOne(repoDoc);
         repository.setId(repoDoc.getObjectId("_id").toString());
 
-        for (File file : repository.getFiles()) {
-            file.setRepoId(repository.getId());
-            saveFile(file);
-        }
-
         return repository;
     }
 
-    public File saveFile(File file) {
+    public File saveFile(File file, String email) {
+    	file.setEmail(email);
         Document fileDoc = convertFileToDocument(file);
         MongoCollection<Document> fileCollection = getFileCollection();
         fileCollection.insertOne(fileDoc);
@@ -70,23 +69,39 @@ public class MongoRepoService {
         return file;
     }
 
-    public void deleteFile(String fileId) {
+    public void deleteFile(String fileId, String email) {
         MongoCollection<Document> fileCollection = getFileCollection();
-        fileCollection.deleteOne(eq("_id", new ObjectId(fileId)));
+        Document fileDoc = fileCollection.find(eq("_id", new ObjectId(fileId))).first();
+        if (fileDoc != null) {
+            if (!fileDoc.getString("email").equals(email)) {
+                throw new RuntimeException("Unauthorized");
+            }
+            fileCollection.deleteOne(eq("_id", new ObjectId(fileId)));
+        } else {
+            throw new RuntimeException("File not found");
+        }
     }
 
-    public Repository updateRepository(String repoId, Repository updatedRepository) {
-        
-        Document updateDoc = new Document()
-                .append("repoName", updatedRepository.getRepoName())
-                .append("repoDescription", updatedRepository.getRepoDescription())
-                .append("repoTopicTags", updatedRepository.getRepoTopicTags())
-                .append("isPublic", updatedRepository.isPublic());
-
+    public Repository updateRepository(String repoId, Repository updatedRepository, String email) {
         MongoCollection<Document> repoCollection = getRepoCollection();
-        repoCollection.updateOne(eq("_id", new ObjectId(repoId)), new Document("$set", updateDoc));
-        updatedRepository.setId(repoId);
-        return updatedRepository;
+        Document repoDoc = repoCollection.find(eq("_id", new ObjectId(repoId))).first();
+        if (repoDoc != null) {
+            if (!repoDoc.getString("email").equals(email)) {
+                throw new RuntimeException("Unauthorized");
+            }
+
+            Document updateDoc = new Document()
+                    .append("repoName", updatedRepository.getRepoName())
+                    .append("repoDescription", updatedRepository.getRepoDescription())
+                    .append("repoTopicTags", updatedRepository.getRepoTopicTags())
+                    .append("isPublic", updatedRepository.isPublic());
+
+            repoCollection.updateOne(eq("_id", new ObjectId(repoId)), new Document("$set", updateDoc));
+            updatedRepository.setId(repoId);
+            return updatedRepository;
+        } else {
+            throw new RuntimeException("Repository not found");
+        }
     }
 
     public List<Repository> getAllPublicRepositories() {
@@ -125,10 +140,19 @@ public class MongoRepoService {
         }
     }
 
-    public void deleteRepositoryById(String repoId) {
+    public void deleteRepositoryById(String repoId, String email) {
         MongoCollection<Document> repoCollection = getRepoCollection();
-        repoCollection.deleteOne(eq("_id", new ObjectId(repoId)));
+        Document repoDoc = repoCollection.find(eq("_id", new ObjectId(repoId))).first();
+        if (repoDoc != null) {
+            if (!repoDoc.getString("email").equals(email)) {
+                throw new RuntimeException("Unauthorized");
+            }
+            repoCollection.deleteOne(eq("_id", new ObjectId(repoId)));
+        } else {
+            throw new RuntimeException("Repository not found");
+        }
     }
+
 
     private Repository convertDocumentToRepository(Document doc) {
         Repository repository = new Repository();
@@ -138,20 +162,7 @@ public class MongoRepoService {
         repository.setRepoDescription(doc.getString("repoDescription"));
         repository.setRepoTopicTags(doc.getList("repoTopicTags", String.class));
         repository.setPublic(doc.getBoolean("isPublic"));
-
-        List<File> files = new ArrayList<>();
-        if (doc.get("files") != null) {
-            for (Document fileDoc : (List<Document>) doc.get("files")) {
-                File file = new File();
-                file.setId(fileDoc.getObjectId("_id").toString());
-                file.setRepoId(fileDoc.getString("repoId")); // Set the repoId
-                file.setFileName(fileDoc.getString("fileName"));
-                file.setFileContent(fileDoc.getString("fileContent"));
-                files.add(file);
-            }
-        }
-        repository.setFiles(files);
-
+        repository.setEmail(doc.getString("email"));
         return repository;
     }
 }
